+++ lib/kitchen/document.rb
@@ -34,7 +34,10 @@ module Kitchen
     # @!method to_html
     #   @see https://www.rubydoc.info/github/sparklemotion/nokogiri/Nokogiri/XML/Node#to_html-instance_method Nokogiri::XML::Node#to_html
     #   @return [String] the document as an HTML string
-    def_delegators :@nokogiri_document, :to_xhtml, :to_s, :to_xml, :to_html
+    # @!method encoding
+    #   @see https://www.rubydoc.info/github/sparklemotion/nokogiri/Nokogiri/XML/Document#encoding-instance_method Nokogiri::XML::Document#encoding
+    #   @return [String] the document as an HTML string
+    def_delegators :@nokogiri_document, :to_xhtml, :to_s, :to_xml, :to_html, :encoding
 
     # Return a new instance of Document
     #
@@ -186,6 +189,17 @@ module Kitchen
     #
     def raw
       @nokogiri_document
+    end
+
+    # Returns the locale for this document, default to `:en` if no locale detected
+    #
+    # @return [Symbol]
+    #
+    def locale
+      raw.root['lang']&.to_sym || begin
+        warn 'No `lang` attribute on this document so cannot detect its locale; defaulting to `:en`'
+        :en
+      end
     end
 
     protected