+++ lib/kitchen/oven.rb
@@ -28,6 +28,8 @@ module Kitchen
         config: config
       )
 
+      I18n.locale = doc.locale
+
       [recipes].flatten.each do |recipe|
         recipe.document = doc
         recipe.bake
@@ -35,7 +37,7 @@ module Kitchen
       profile.baked!
 
       File.open(output_file, 'w') do |f|
-        f.write doc.to_xhtml(indent: 2)
+        f.write doc.to_xhtml(indent: 2, encoding: doc.encoding || 'utf-8')
       end
       profile.written!